Tidy Up and Profile Header

+++ services/web/api/route/image.py
@@ -1,3 +1,4 @@
+from crypt import methods
 from flask import (
     Blueprint,
     jsonify,
@@ -6,6 +7,7 @@ from flask import (
 )
 from werkzeug.utils import secure_filename
 from api.model.data_spec import Image
+from api.helper.auth import token_required
 from .. import db
 from pathlib import Path
 
@@ -25,7 +27,7 @@ def getImage():
     if 'imageid' not in request.args:
         return jsonify(message="Please supply imageid"), 422
     
-    return jsonify(message=f"Returing user with ID {request.args['imageid']}", value=Image.query.filter_by(image_id=request.args['imageid']).first()), 200
+    return jsonify(message=f"Returing image with ID {request.args['imageid']}", value=Image.query.filter_by(image_id=request.args['imageid']).first()), 200
 
 @image.route("/author")
 def imagesByAuthor():
@@ -42,17 +44,12 @@ def imagesByGallery():
     return jsonify(value=Image.query.filter_by(gallery_id=request.args['galleryid']).all()), 200
 
 @image.route("/generate", methods=["POST"])
-def generateImage():
-    data = request.get_json()
-    owner_id = data['owner_id'] if 'owner_id' in data else 0
-    
-    if owner_id == 0:
-        current_app.logger.debug(data)
-        return jsonify(message="Please supply owner_id"), 422
+@token_required
+def generateImage(current_user):
+    owner_id = current_user.user_id
     
     dir_name = hashlib.md5(str(owner_id).encode())
     Path(f"api/images/{dir_name.hexdigest()}").mkdir(parents=True, exist_ok=True)
-    current_app.logger.debug(data)
     image_id = uuid.uuid4()
     image = random.choice(os.listdir("./api/images/test/"))
     new_image = Image(image_id=image_id, owner_id=owner_id, file_location=f"{image}", name=str(uuid.uuid4()))
@@ -63,7 +60,8 @@ def generateImage():
     return jsonify(message="Image Created", value=new_image)
 
 @image.route("/modify", methods=["PATCH"])
-def modifyImage():
+@token_required
+def modifyImage(current_user):
     if not request.is_json:
         return jsonify(message="Missing JSON in request"), 400
     
@@ -86,19 +84,19 @@ def modifyImage():
     cur_image.description = data["description"] if "description" in data else cur_image.description
     
     current_app.logger.debug(f"After update: {cur_image}")
+    current_app.logger.info(f"User {current_user.user_id} updated image with ID {cur_image.image_id}")
     
     db.session.commit()
     
     return jsonify(message=f"Returing image with ID {image_id}", value=cur_image), 200
 
 @image.route("/upload", methods=["POST"])
-def imageUpload():
+@token_required
+def imageUpload(current_user):
     if 'file' not in request.files:
         return jsonify(message="Missing file data in request"), 400
-    if 'owner' not in request.args:
-        return jsonify(message="Missing owner arg in request"), 400
     
-    dir_name = hashlib.md5(str(request.args['owner']).encode())
+    dir_name = hashlib.md5(str(current_user.user_id).encode())
     Path(f"api/images/uploaded/{dir_name.hexdigest()}").mkdir(parents=True, exist_ok=True)
     
     uploaded_file = request.files['file']
@@ -111,19 +109,55 @@ def imageUpload():
         image_id = uuid.uuid4()
         
         uploaded_file.save(f"api/images/uploaded/{dir_name.hexdigest()}/{image_id}{file_ext}")
-        new_image = Image(image_id=image_id, owner_id=request.args['owner'], file_location=f"/images/{dir_name.hexdigest()}/{image_id}{file_ext}", name=str(uuid.uuid4()))
+        new_image = Image(image_id=image_id, owner_id=current_user.user_id, file_location=f"/images/uploaded/{dir_name.hexdigest()}/{image_id}{file_ext}", name=str(uuid.uuid4()))
     
+    current_app.logger.info(f"User {current_user.user_id} uploaded image with ID {new_image.image_id}")
+
     db.session.add(new_image)
     db.session.commit()
 
     return jsonify(message="Image Uploaded", value=new_image)
 
+@image.route("/save", methods=["POST"])
+@token_required
+def saveImage(current_user):
+    if 'file' not in request.files:
+        return jsonify(message="Missing file data in request"), 400
+    
+    dir_name = hashlib.md5(str(current_user.user_id).encode())
+    Path(f"api/images/combined/{dir_name.hexdigest()}").mkdir(parents=True, exist_ok=True)
+    
+    uploaded_file = request.files['file']
+    filename = secure_filename(uploaded_file.filename)
+    if filename != '':
+        file_ext = os.path.splitext(filename)[1]
+        if file_ext not in current_app.config['UPLOAD_EXTENSIONS']:
+            return jsonify(message="Image not accepted file type"), 400
+        
+        image_id = uuid.uuid4()
+        
+        uploaded_file.save(f"api/images/combined/{dir_name.hexdigest()}/{image_id}{file_ext}")
+        new_image = Image(image_id=image_id, owner_id=current_user.user_id, file_location=f"/images/combined/{dir_name.hexdigest()}/{image_id}{file_ext}", name=str(uuid.uuid4()))
+    
+    current_app.logger.info(f"User {current_user.user_id} saved image with ID {new_image.image_id}")
+    
+    db.session.add(new_image)
+    db.session.commit()
+
+    return jsonify(message="Image Saved", value=new_image)
+
 @image.route("/delete", methods=["DELETE"])
-def deleteImage():
+@token_required
+def deleteImage(current_user):
     if 'imageid' not in request.args:
         return jsonify(message="Missing owner arg in request"), 400
     
     cur_img = Image.query.filter_by(image_id=request.args['imageid']).first()
+
+    if not current_user.user_id == cur_img.owner_id:
+        return jsonify(message="You must own this image to delete it")
+    
+    current_app.logger.info(f"User {current_user.user_id} deleted image with ID {cur_img.image_id}")
 
     db.session.delete(cur_img)
     db.session.commit()
